+++ OneBlink.SDK/models/FormSubmissionEventCivicaRecord.cs
@@ -1,0 +1,8 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormSubmissionEventCivicaRecord
+    {
+        public long id { get; set; }
+        public string label { get; set; }
+    }
+}